refactor(pie-label): 饼图label的调整，移动至layout中处理（需要将labelItem传递至layoutFn）
- [x] 重构
- [ ] 测例
